+++ HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+1.0.0-alpha0 (2024-07-12)
++++++++++++++++++++
+* Use generic azure token credential for auth instead of custom lib.auth
+* Remove older Python support
+
 0.0.53 (2023-04-11)
 +++++++++++++++++++
 * Add MSAL support. Remove ADAL support